feat: add class to fetch and display weather
This commit introduces a new Weather class that extends HomeDashboard. It fetches current
and forecasted weather data from WeatherAPI, and displays it on an LCD. It also manages data
storage to minimize API requests, and handles asynchronous updates to keep the display current.
The data is displayed with a timestamp indicating the last update, and there's a distinction
between displaying newly fetched data and previously stored data.

- One view: Displays temperature and condition for current day. Updated every 15 minutes.
- Second view: Displays temp and condition for tomorrow. Updated every hour.

TO-DO:
- Introduce better time tracking in 1st view.
- Create long string display functionality. (scrolling text?) Some conditions, like "Patchy rain possible" are longer than 16 characters and thus muff up the display.
- Add different data views for different verbosity levels.
- Add functionality to switch btwn "views"