fix(btn): fixed view cycling
Before this change, cycling through screens on the LCD wasn't working as intended. I believe this problem arises due to how the functions inside of their respective classes updated the LCD display. For example, I originally had a function called **poll_date_time** inside of the **DateTime** class which used a while True loop to continuously update the date/time and print this information to the LCD display. In the future, I think the most optimal way will be to use a while loop inside of the **HomeDashboard** class which calls **poll_date_time** continuously. The **poll_date_time** function therefore will only do one polling. It's essentially the same functionality. However, instead of the while loop being inside of the nested function, the while loop will now continuously call the function.

After this change, I've moved from actually displaying the various screens to just a template. The actual screens have to be looked at once again, as some of the data fetching is messing with how the screens are updated. Still, the buttons now properly cycle through the 4 views, with an alternate screen for each view that can be accessed using the secondary button.

I've also changed the while True loop to a while loop with a flag variable. This allows greater control over the while loop. For more detail, check the docstring for the **button_pressed_callback** function inside of the **HomeDashboard** class.

TO-DO: Currently, the "views" are just template messages (e.g. "date/time view", "weather view", and "alt view"). These will need to be updated so that they are actually showing the view with real data.