feat: add dinner view
- Added dinner_data.json file to store planned dinners
- Planned dinners are structured: week -> day -> main course & sides 1, 2, & 3
- Added `get_main_course` function which takes args date (MM/DD) and day_of_week (monday, tuesday, etc.) and pulls the main course
- Added `get_sides` function which takes args date and day_of_week and pulls the sides into a list
- `get_sides` functions ignores sides stored as None
- Added display functions to print this data to the LCD display
- The length of the string will determine whether or not to center the string on the display or use scrolling text
- The list of sides will be converted to a string with plus signs (+) connecting the sides (i.e. ['peppers', 'onions'] becomes "Sides: peppers + onions"

TO-DO: Create logic which gets the current day of the week. The logic will need to compare the current date to the monday of that week to get to the proper week inside of the json file. Once the date of the monday is found, and the current day of the week is known, we can call the get_main_course and get_sides functions and print these to the LCD display.
TO-DO: Add an alternate view which shows tomorrow's dinner plans
TO-DO: Add an alternate view which shows a random dinner plan