feat(web interface): updated webpage for sending messages
This commit features an overhaul of the web interface used to send messages to the LCD display over a webpage. I have moved the Flask application into a class which simplifies importing the app into the msg_view file as well as simplifying how the message is transferred to the LCD display. For example, the "message" is now stored as an attribute inside of WebApp and can be accessed simply by using "{web_app_instance}.message". This commit also adds a CSS file for styling the webpage, as well as a homepage (index.html) for future improvements to the web interface.

- renamed _app.py_ to _web_interface.py_
- encapsulated logic into a class: WebApp
- added seperate pages for the homepage (index.html) and the text submission page (submit_msg.html)
- added CSS file (styles.css) that is used by Flask application
- updated msg_view.py by adding **msg_display** function
- message display function access the WebApp instance to get the message and writes it to LCD display
- TO-DO: update msg_view.py
- TO-DO: create WebApp instance in a seperate thread so that main controller functionality can still run
- TO-DO: add logic to **msg_display** function which can handle long messages, and lines 1 and 2...