fix: weather view
Weather view has undergone a complete overhaul. I've switched APIs and landed on one which is free, open-source, and requires no key. This makes it easier for others who want to use the project, as they no longer have to go and sign up for an account somewhere and get their own key. I've also changed up how I fetch data. BEforehand, I was using the asyncio library and the aiohttp library. Admittedly, I didn't fully understand how these worked. And I've come to find out, I didn't really need to use these libraries. Asyncio and aiohttp can be really powerful, especially in the case of making many API requests (100s to 1000s in minutes etc.). However, we really only need to make one API request every 15 minutes. This is because the weather API updates data on their end every 15 minutes.

I've also improved my understanding of fetching from an API in general, which has led to a cleaner 'fetch_weather' function. Now, a base url is stored, and a dictionary of parameters are stored. These parameters are passed into the requests.get function, along with the base URL, to get the data necessary for our displays. This makes expanding in the future a lot easier, as we only have to go and update the params dict as opposed to working within a long url string.

I should have mentioned this earlier, but a Lock is used with the multithreaded approach to ensure that the file manipulation occuring by the multiple threads is done at different times.

This fix:
- Switches to a free, open-source API with no key authentication
- Changes approach from asyncio and aiohttp to multiple threads with a lock. Main thread handles display and extracting data from file. Background thread fetches data from API every 10 minutes and updates the file.

NOTE: When the program is first launched, it's very likely that the weather data will be outdated. In this case, I think it'd be best to present a loading screen to the LCD display, while we wait for the new data. This is only a problem that arises on FIRST launching the software. Because, otherwise, the new data will be fetched every 10 minutes.